added minor changes in wykopproducer/app.py deleted wyko_client_mock package -> wykop_client_mock.py is now part of wykop_client.py
